:recycle: Declouped and improved deficiencies creation flow
